llama/requirements.txt
google.generativeai
streamlit==1.30.0